cleanCodeAndCleanArchitecture
