+++ package.json
@@ -4,30 +4,30 @@
   "version": "0.1.0",
   "private": false,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/js-cookie": "^3.0.3",
-    "@xstate/react": "^3.2.2",
-    "autoprefixer": "^9.8.6",
-    "cookies-ts": "^1.0.5",
-    "date-fns": "^2.30.0",
-    "js-cookie": "^3.0.5",
-    "mobx": "^6.10.1",
-    "mobx-react": "^9.0.1",
-    "mobx-react-lite": "^4.0.4",
-    "postcss": "^8.4.29",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.10.1",
-    "react-loader-spinner": "^5.3.4",
+    "@testing-library/jest-dom": "5.17.0",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "13.5.0",
+    "@types/js-cookie": "3.0.3",
+    "@xstate/react": "3.2.2",
+    "autoprefixer": "9.8.6",
+    "cookies-ts": "1.0.5",
+    "date-fns": "2.30.0",
+    "js-cookie": "3.0.5",
+    "mobx": "6.10.1",
+    "mobx-react": "9.0.1",
+    "mobx-react-lite": "4.0.4",
+    "postcss": "8.4.29",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-icons": "4.10.1",
+    "react-loader-spinner": "5.3.4",
     "react-router-dom": "6.14.2",
-    "react-scripts": "^5.0.1",
-    "reactjs-popup": "^2.0.5",
-    "recharts": "^2.7.2",
-    "uuid": "^9.0.0",
-    "web-vitals": "^2.1.4",
-    "xstate": "^4.38.2"
+    "react-scripts": "5.0.1",
+    "reactjs-popup": "2.0.5",
+    "recharts": "2.7.2",
+    "uuid": "9.0.0",
+    "web-vitals": "2.1.4",
+    "xstate": "4.38.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -56,7 +56,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^5.0.0",
-    "tailwindcss": "^3.3.3"
+    "gh-pages": "5.0.0",
+    "tailwindcss": "3.3.3"
   }
 }
