+++ src/App.tsx
@@ -1,11 +1,12 @@
-import {Routes, Route, Navigate} from 'react-router-dom'
+import {Routes, Route} from 'react-router-dom'
+import { LOG_IN_PATH, HOME_PATH, TRANSACTIONS_PATH, PROFILE_PATH, NOT_FOUND_PATH } from './constants/NavigationConstants'
 
-import LoginPage from './components/LoginPage'
-import Home from './components/Home'
-import Transactions from './components/Transactions'
-import Profile from './components/Profile'
-import NotFound from './components/NotFound'
-import ProtectedRoute from './components/ProtectedRoute'
+import LoginPage from './components/LoginPage/LoginPage'
+import Home from './components/Home/Home'
+import Transactions from './components/Transactions/Transactions'
+import Profile from './components/Profile/Profile'
+import NotFound from './components/NotFound/NotFound'
+import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute'
 import {StoreProvider}  from './context/storeContext'
 
 import './App.css'
@@ -14,13 +15,13 @@ const App = () => {
   return(
     <StoreProvider>
       <Routes>
-        <Route path="/login" element={<LoginPage/>} />
+        <Route path={LOG_IN_PATH} element={<LoginPage/>} />
         <Route element={<ProtectedRoute/>}>
-          <Route path="/" element={<Home/>} />
-          <Route  path="/transactions" element={<Transactions/>}/>
-          <Route  path="/profile" element={<Profile/>}/>
+          <Route path={HOME_PATH} element={<Home/>} />
+          <Route  path={TRANSACTIONS_PATH} element={<Transactions/>}/>
+          <Route  path={PROFILE_PATH} element={<Profile/>}/>
         </Route>
-        <Route path="/not-found" element={<NotFound/>} />
+        <Route path={NOT_FOUND_PATH} element={<NotFound/>} />
       </Routes>
     </StoreProvider>
   )
