+++ src/components/NotFound/index.tsx
@@ -1,13 +1,3 @@
-import './index.css'
-
-const NotFound = () => (
-  <div className="not-found-container">
-    <img
-      src="https://assets.ccbp.in/frontend/react-js/not-found-blog-img.png"
-      alt="not found"
-      className="not-found-img"
-    />
-  </div>
-)
+import NotFound from "./NotFound";
 
 export default NotFound
