+++ src/components/ProtectedRoute/index.tsx
@@ -1,10 +1,3 @@
-import { Navigate, Outlet} from 'react-router-dom'
+import ProtectedRoute from "./ProtectedRoute";
 
-import useUserId from '../../hooks/getUserId'
-
-
-const ProtectedRoute = () => {
-  const userId = useUserId()
-  return userId === undefined ? <Navigate to ="/login"/> : <Outlet />
-}
 export default ProtectedRoute
