+++ src/components/TabItem/index.css
@@ -18,6 +18,6 @@
     font-size: 20px;
     padding-bottom: 2px;
 }
-.tabitem{
+.tab-item{
     list-style: none;
 }
