+++ tailwind.config.js
@@ -307,6 +307,7 @@ module.exports = {
         '42px': '42px',
         '44px': '44px',
         '50px': '50px',
+        '58px': '58px',
         '60px': '60px',
         '80px': '80px',
         '160px': '160px',
@@ -321,6 +322,7 @@ module.exports = {
         '345px': '345px',
         '282px': '282px',
         '1100px': '1100px',
+        '1061px': '1061px',
         '80vh': '80vh',
         '60vh': '60vh',
         '70vw': '70vw',
